requirements.txt
PyYAML
requests